requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
Django==2.2
django-cors-headers==3.7.0
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
